requirements.txt
# Python clothing prediction service (clothing_predict_server.py)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pillow>=9.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0

# Optional for better clothing type prediction (CPU-only, separate install):
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
